Add new files and update existing ones for API improvements
Implemented the addition of new files and updates to existing ones to enhance the functionality of the BudgetPlanner API. This includes the introduction of fresh classes and methods, an augmentation to the database context, and modifications to the API Bootstrap. The Newtonsoft.Json package was also added to the BudgetPlaner.Models project file.