Refactor domain models and update endpoints
Improved the clarity and consistency of the data models and respective endpoints for the BudgetPlaner application. Deleted the CreditInterestRate model as it was no longer needed, renamed several classes to better match their function, and made numerous adjustments to properties and relationships for improved data integrity. Also added error handling and user validation to many endpoint actions for improved security and usability. Introduced additional FluentProvider to enhance UI.